requirements.txt
tensorflow>=2.8.0
gradio>=3.0.0
pillow>=8.0.0
numpy>=1.19.0